Create an App.js For The Server.
Perhaps somewhat contentious but, this allows the app object to be
pulled into all required files and keeps its' instantiation in a more
'top level' space. This would theoretically allow swapping of the server
without having to change the index or any route specific files.

Any other changes are purely aesthetic and due to linting.